# openapi.yaml
openapi: 3.0.3
info:
  title: LeanStock Backend API
  description: |
    Production-grade inventory management system for coffee equipment retailers.
    Handles multi-tenant stock tracking, dead-stock decay pricing, demand forecasting,
    and full audit logging across warehouses.
  version: 1.0.0
  contact:
    name: LeanStock
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Pagination:
      type: object
      properties:
        nextCursor:
          type: string
          nullable: true
          example: "eyJpZCI6ImFiYzEyMyJ9"
        hasMore:
          type: boolean
          example: true
        limit:
          type: integer
          example: 20
...
    Role:
      type: string
      enum: [ADMIN, MANAGER, STOREKEEPER]
    Category:
      type: string
      enum: [EQUIPMENT, COFFEE, ACCESSORIES]
    ProductStatus:
      type: string
      enum: [ACTIVE, DISPLAY_SAMPLE, DISCONTINUED, CRITICAL_DEADSTOCK]
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        email:
          type: string
          format: email
          example: "manager@coffeeshop.kz"
        role:
          $ref: '#/components/schemas/Role'
        isActive:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
...
    Product:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "b1c2d3e4-f5a6-7890-abcd-ef1234567890"
        sku:
          type: string
          example: "JURA-E8-BLK"
        name:
          type: string
          example: "Jura E8 Black"
        brand:
          type: string
          example: "Jura"
        category:
          $ref: '#/components/schemas/Category'
        status:
          $ref: '#/components/schemas/ProductStatus'
        basePrice:
          type: integer
          description: Original price in KZT (tenge)
          example: 285000
        costPrice:
          type: integer
          description: Procurement cost in KZT
          example: 190000
        currentPrice:
          type: integer
          description: Effective selling price after any decay discount
          example: 256500
        discountPercent:
          type: integer
          description: Applied decay discount (0-50%)
          example: 10
        daysInInventory:
          type: integer
          description: Days since last stock replenishment
          example: 45
        tenantId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Stock:
      type: object
      properties:
        id:
          type: string
          format: uuid
        quantity:
          type: integer
          example: 7
        reorderThreshold:
          type: integer
          example: 5
        productId:
          type: string
          format: uuid
        locationId:
          type: string
          format: uuid
        updatedAt:
          type: string
          format: date-time
...
    Location:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "d4e5f6a7-b8c9-0123-def0-1234567890ab"
        name:
          type: string
          example: "Almaty Warehouse"
        address:
          type: string
          example: "Abay Ave 150, Almaty"
        tenantId:
          type: string
          format: uuid
...
    ForecastResult:
      type: object
      properties:
        product:
          type: object
          properties:
            id:
              type: string
              format: uuid
            sku:
              type: string
            name:
              type: string
        forecast:
          type: object
          properties:
            periodWeeks:
              type: integer
              example: 4
            totalSoldInPeriod:
              type: integer
              example: 28
            avgWeeklySales:
              type: number
              format: float
              example: 7.0
            reorderPoint:
              type: integer
              description: Units (2-week lead time × avg weekly sales)
              example: 14
            currentStock:
              type: integer
              example: 7
            weeksOfStockLeft:
              type: string
              example: "1.0"
            needsReorder:
              type: boolean
              example: true
        stockByLocation:
          type: array
          items:
            type: object
            properties:
              location:
                type: string
              quantity:
                type: integer
paths:
...
  /users:
    get:
      tags: [Users]
      summary: List all staff in tenant (ADMIN)
      operationId: listUsers
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: cursor
          in: query
          schema:
            type: string
          description: Base64-encoded pagination cursor
      responses:
        '200':
          description: Paginated staff list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          description: Not authenticated
        '403':
          description: Not ADMIN

    post:
      tags: [Users]
      summary: Create new staff account (ADMIN)
      operationId: createUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  example: "newstaff@coffeeshop.kz"
                password:
                  type: string
                  minLength: 8
                  example: "StaffPass1"
                role:
                  $ref: '#/components/schemas/Role'
      responses:
        '201':
          description: Staff account created
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
        '409':
          description: Email already registered
  /users/{id}:
    get:
      tags: [Users]
      summary: Get single staff member (ADMIN)
      operationId: getUser
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: Not found

    patch:
      tags: [Users]
      summary: Update role or active status (ADMIN)
      operationId: updateUser
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                role:
                  $ref: '#/components/schemas/Role'
                isActive:
                  type: boolean
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '403':
          description: Cannot modify your own account
...
  /locations:
    get:
      tags: [Locations]
      summary: List all locations in tenant
      operationId: listLocations
      security:
        - bearerAuth: []
      responses:
        '200':
          description: All locations
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Location'

    post:
      tags: [Locations]
      summary: Create a new warehouse/store location (ADMIN)
      operationId: createLocation
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, address]
              properties:
                name:
                  type: string
                  example: "Almaty Warehouse"
                address:
                  type: string
                  example: "Abay Ave 150, Almaty"
      responses:
        '201':
          description: Location created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Location'
  /locations/{id}:
    get:
      tags: [Locations]
      summary: Get location with current stock snapshot
      operationId: getLocation
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Location with stock details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Location'
                  - type: object
                    properties:
                      stocks:
                        type: array
                        items:
                          allOf:
                            - $ref: '#/components/schemas/Stock'
                            - type: object
                              properties:
                                product:
                                  $ref: '#/components/schemas/Product'
        '404':
          description: Location not found
  /forecast:
    get:
      tags: [Forecast]
      summary: Demand forecast for all active products (MANAGER+)
      operationId: getAllForecasts
      security:
        - bearerAuth: []
      parameters:
        - name: weeks
          in: query
          schema:
            type: integer
            minimum: 1
            default: 4
          description: Number of weeks of sales history to analyze
      responses:
        '200':
          description: Forecasts sorted by urgency (needsReorder first)
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ForecastResult'
                  weeksAnalyzed:
                    type: integer
                    example: 4
  /forecast/{productId}:
    get:
      tags: [Forecast]
      summary: Demand forecast for a single product (MANAGER+)
      operationId: getProductForecast
      security:
        - bearerAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: weeks
          in: query
          schema:
            type: integer
            minimum: 1
            default: 4
      responses:
        '200':
          description: Product forecast
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForecastResult'
        '404':
          description: Product not found
...